add json option
