+++ lib/prest/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Prest
-  VERSION = '0.1.5'
+  VERSION = '0.1.6'
 end
